cli: split cli.go into multiple files
